refactor int
