+++ cmd/tracko/main.go
@@ -8,6 +8,8 @@ import (
 
 	"github.com/tendermint/basecoin/cmd/commands"
 	"github.com/tendermint/tmlibs/cli"
+
+	_ "github.com/tendermint/trackomatron/cmd/tracko/commands"
 	_ "github.com/tendermint/trackomatron/commands"
 )
